requirements.txt
altgraph==0.17.5
asgiref==3.11.1
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
django-rest-framework==0.1.0
djangorestframework==3.17.1
python-dotenv==1.0.1
iniconfig==2.3.0
packaging==26.0
pefile==2024.8.26
pillow==12.1.1
pluggy==1.6.0
Pygments==2.20.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
PyJWT==2.12.1
pytest==9.0.2
pywin32-ctypes==0.2.3
setuptools==82.0.1
sqlparse==0.5.5
tzdata==2025.3
waitress==3.0.2
